CS dialogue @ Tanamor + Optimize naming system
I added an inner dialogue for the MC when approaching Tanamor Tower. In addition, I found a text command I can use to fitch the name of the characters through Actor ID. That way I don't have to change the name of the character on each text dialogue, I just need to change it from the actor menu in the Database center.